sep the file and functiongit add src/parser/!
