+++ src/parser/ft_isspace.c
@@ -1,0 +1,10 @@
+#include "parser.h"
+
+int	ft_isspace(char c)
+{
+	if (c == '\r' || c == '\n' || c == '\f' )
+		return (1);
+	else if (c == '\v' || c == '\t' || c == ' ' )
+		return (1);
+	return (0);
+}
