+++ src/parser/init_map.c
@@ -1,0 +1,19 @@
+#include "parser.h"
+
+char	**init_map(char **smap, int width, int height)
+{
+	char	**map;
+	int		i;
+
+	i = 0;
+	(void)width;
+	map = (char **)malloc(sizeof(char *) * (height + 1));
+	map[height] = NULL;
+	while (i < height)
+	{
+		map[i] = ft_strdup(smap[i + 6]);
+		i++;
+	}
+	free2d(smap);
+	return (map);
+}
